// src/models/Ide.ts
export class Ide {
  cUF: string;
  cNF: string;
  natOp: string;
  indPag: string;
  mod: string;
  serie: string;
  nNF: string;
  dEmi: string;
  dSaiEnt: string;
  tpNF: string;
  cMunFG: string;
  tpImp: string;
  tpEmis: string;
  cDV: string;
  tpAmb: string;
  finNFe: string;
  procEmi: string;
  verProc: string;
}