// src/models/Produto.ts
import { CEan } from "./CEan";
import { CEantrib } from "./CEantrib";

export class Produto {
  cProd: string;
  cEAN: CEan;
  xProd: string;
  CFOP: string;
  uCom: string;
  qCom: string;
  vUnCom: string;
  vProd: string;
  cEANTrib: CEantrib;
  uTrib: string;
  qTrib: string;
  vUnTrib: string;
}